+++ Tests/ChessKitTests/GameTests.swift
@@ -12,12 +12,12 @@ class GameTests: XCTestCase {
 
     // MARK: - Indices used in tests
 
-    private let nf3Index = MoveTreeIndex(number: 2, color: .white, variation: 0)
-    private let nc3Index = MoveTreeIndex(number: 2, color: .white, variation: 1)
-    private let nf6Index = MoveTreeIndex(number: 2, color: .black, variation: 1)
-    private let nc6Index = MoveTreeIndex(number: 2, color: .black, variation: 2)
-    private let nc6Index2 = MoveTreeIndex(number: 2, color: .black, variation: 0)
-    private let f5Index = MoveTreeIndex(number: 2, color: .black, variation: 3)
+    private let nf3Index = MoveTree.Index(number: 2, color: .white, variation: 0)
+    private let nc3Index = MoveTree.Index(number: 2, color: .white, variation: 1)
+    private let nf6Index = MoveTree.Index(number: 2, color: .black, variation: 1)
+    private let nc6Index = MoveTree.Index(number: 2, color: .black, variation: 2)
+    private let nc6Index2 = MoveTree.Index(number: 2, color: .black, variation: 0)
+    private let f5Index = MoveTree.Index(number: 2, color: .black, variation: 3)
 
     // MARK: - Setup
 
@@ -66,11 +66,11 @@ class GameTests: XCTestCase {
         XCTAssertEqual(game2.startingPosition, .init(fen: fen)!)
         game2.make(move: "O-O", from: game2.startingIndex)
         XCTAssertEqual(
-            game2.moves.nextIndex(for: game2.moves.minimumIndex),
+            game2.moves.index(after: game2.moves.minimumIndex),
             .init(number: 1, color: .black, variation: 0)
         )
         XCTAssertEqual(
-            game2.moves.move(at: .init(number: 1, color: .black, variation: 0)),
+            game2.moves[.init(number: 1, color: .black, variation: 0)],
             .init(san: "O-O", position: .init(fen: fen)!)
         )
     }
@@ -92,8 +92,8 @@ class GameTests: XCTestCase {
         XCTAssertEqual(game.moves[nc6Index]?.san, "Nc6")
 
         XCTAssertEqual(
-            game.moves.previousIndex(
-                for: nc6Index
+            game.moves.index(
+                before: nc6Index
             ),
             nc3Index
         )
@@ -102,14 +102,14 @@ class GameTests: XCTestCase {
         XCTAssertEqual(game.moves[f5Index]?.san, "f5")
 
         XCTAssertEqual(
-            game.moves.previousIndex(
-                for: f5Index
+            game.moves.index(
+                before: f5Index
             ),
             nf3Index
         )
 
-        XCTAssertEqual(game.moves.previousIndex(for: .minimum.next), .minimum)
-        XCTAssertEqual(game.moves.nextIndex(for: nc3Index), nf6Index)
+        XCTAssertEqual(game.moves.index(before: .minimum.next), .minimum)
+        XCTAssertEqual(game.moves.index(after: nc3Index), nf6Index)
     }
 
     func testMoveAnnotation() {
@@ -127,12 +127,10 @@ class GameTests: XCTestCase {
 
         XCTAssertEqual(
             f5History,
-            [
-                .init(number: 1, color: .white, variation: 0),
-                .init(number: 1, color: .black, variation: 0),
-                .init(number: 2, color: .white, variation: 0),
-                f5Index
-            ]
+            [.init(number: 1, color: .white, variation: 0),
+             .init(number: 1, color: .black, variation: 0),
+             .init(number: 2, color: .white, variation: 0),
+             f5Index]
         )
     }
 
@@ -141,10 +139,16 @@ class GameTests: XCTestCase {
 
         XCTAssertEqual(
             f5Future,
-            [
-                .init(number: 3, color: .white, variation: 3)
-            ]
+            [.init(number: 3, color: .white, variation: 3)]
         )
+    }
+
+    func testMoveIndexFullVariation() {
+        let f5History = game.moves.history(for: f5Index)
+        let f5Future = game.moves.future(for: f5Index)
+
+        let f5Full = game.moves.fullVariation(for: f5Index)
+        XCTAssertEqual(f5History + f5Future, f5Full)
     }
 
     func testMoveTreeEmptyPath() {
@@ -153,19 +157,19 @@ class GameTests: XCTestCase {
 
     func testMoveTreeSimplePath() {
         // "1. e4 e5 2. Nf3 (2. Nc3 Nf6 (2... Nc6 3. f4) 3. Bc4) Nc6 (2... f5 3. exf5) 3. Bc4"
-        let f4 = MoveTreeIndex(number: 3, color: .white, variation: 2)
-        let e5 = MoveTreeIndex(number: 1, color: .black, variation: 0)
+        let f4 = MoveTree.Index(number: 3, color: .white, variation: 2)
+        let e5 = MoveTree.Index(number: 1, color: .black, variation: 0)
 
         // 3. f4 to 1. e5
         let path1 = game.moves.path(from: f4, to: e5)
 
         XCTAssertEqual(
-            path1.map(\.0),
+            path1.map(\.direction),
             [.reverse, .reverse, .reverse]
         )
 
         XCTAssertEqual(
-            path1.map(\.1),
+            path1.map(\.index),
             [
                 f4,
                 .init(number: 2, color: .black, variation: 2),
@@ -177,12 +181,12 @@ class GameTests: XCTestCase {
         let path2 = game.moves.path(from: e5, to: f4)
 
         XCTAssertEqual(
-            path2.map(\.0),
+            path2.map(\.direction),
             [.forward, .forward, .forward]
         )
 
         XCTAssertEqual(
-            path2.map(\.1),
+            path2.map(\.index),
             [
                 .init(number: 2, color: .white, variation: 1),
                 .init(number: 2, color: .black, variation: 2),
@@ -194,12 +198,12 @@ class GameTests: XCTestCase {
     func testMoveTreeComplexPath() {
         // "1. e4 e5 2. Nf3 (2. Nc3 Nf6 (2... Nc6 3. f4) 3. Bc4) Nc6 (2... f5 3. exf5) 3. Bc4"
         // 3. f4 to 3. Bc4
-        let f4 = MoveTreeIndex(number: 3, color: .white, variation: 2)
-        let Bc4 = MoveTreeIndex(number: 3, color: .white, variation: 0)
+        let f4 = MoveTree.Index(number: 3, color: .white, variation: 2)
+        let Bc4 = MoveTree.Index(number: 3, color: .white, variation: 0)
         let path = game.moves.path(from: f4, to: Bc4)
 
         XCTAssertEqual(
-            path.map(\.0),
+            path.map(\.direction),
             [
                 .reverse,
                 .reverse,
@@ -212,7 +216,7 @@ class GameTests: XCTestCase {
         )
         
         XCTAssertEqual(
-            path.map(\.1),
+            path.map(\.index),
             [
                 f4,
                 .init(number: 2, color: .black, variation: 2),